services/data_synthesizer/requirements.txt
google-cloud-pubsub==2.13.0
google-cloud-aiplatform==1.38.1
Faker==22.4.0